+++ puppet/CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Environment group feature for node groups that assign environments instead of filtering by them
+
 ## [0.22.0] - 2025-01-15
 
 ### Added